EL-3268 - Adding Date Picker Localization Support
