+++ docs/app/pages/components/components-sections/date-time-picker/date-time-picker/date-time-picker.component.html
@@ -115,6 +115,48 @@
     </tr>
 </uxd-api-properties>
 
+<p>
+    To provide default settings for all date pickers within you application, import <code>DateTimePickerModule.forRoot()</code> in your <code>app.module.ts</code>.
+    You can then inject the <code>DateTimePickerConfig</code> and configure the default presentation and localization settings for every date picker
+    within your application.
+</p>
+
+<uxd-api-properties tableTitle="DateTimePickerConfig">
+    <tr uxd-api-property name="showDate" type="boolean">
+        Defines whether or not the date picker should be visible.
+    </tr>
+    <tr uxd-api-property name="showTime" type="boolean">
+        Defines whether or not the time picker should be visible.
+    </tr>
+    <tr uxd-api-property name="showTimezone" type="boolean">
+        Defines whether or not the time picker should allow the user to choose a timezone.
+    </tr>
+    <tr uxd-api-property name="showSeconds" type="boolean">
+        Defines whether or not the time picker should allow the user to specify seconds.
+    </tr>
+    <tr uxd-api-property name="showMeridian" type="boolean">
+        Defines whether or not the time picker should show an AM/PM button, or time should be represented in 24hr format instead.
+    </tr>
+    <tr uxd-api-property name="showSpinners" type="boolean">
+        Defines whether or not the time picker should allow the user to select the time using spinners.
+    </tr>
+    <tr uxd-api-property name="weekdays" type="string[]">
+        If defined will override the weekday names displayed.
+    </tr>
+    <tr uxd-api-property name="nowBtnText" type="string">
+        Defines the text to be displayed in the button used to set the selected time to the current time.
+    </tr>
+    <tr uxd-api-property name="timezones" type="DateTimePickerTimezone[]">
+        Defines the timezones that can be selected.
+    </tr>
+    <tr uxd-api-property name="months" type="string[]">
+        Defines the names of the months. This can be used for localization purposes.
+    </tr>
+    <tr uxd-api-property name="monthsShort" type="string[]">
+        Defines the short names of each month. This can be used for localization purposes.
+    </tr>
+</uxd-api-properties>
+
 <p>The following code can be used to create the example above:</p>
 
 <ux-tabset [minimal]="false">
