+++ src/components/date-time-picker/date-time-picker.utils.ts
@@ -1,6 +1,3 @@
-import { DateTimePickerTimezone } from './date-time-picker.service';
-
-
 /**
  * Convert a single dimension array to a double dimension array
  * @param items the single dimension array to convert
@@ -93,3 +90,36 @@ export const monthsShort = months.map(month => month.substring(0, 3));
  */
 export const weekdays = ['Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday', 'Saturday', 'Sunday'];
 export const weekdaysShort = weekdays.map(weekday => weekday.substring(0, 3));
+
+/** Export the default set of time zone */
+export const timezones: DateTimePickerTimezone[] = [
+    { name: 'GMT-11', offset: 660 },
+    { name: 'GMT-10', offset: 600 },
+    { name: 'GMT-9', offset: 540 },
+    { name: 'GMT-8', offset: 480 },
+    { name: 'GMT-7', offset: 420 },
+    { name: 'GMT-6', offset: 360 },
+    { name: 'GMT-5', offset: 300 },
+    { name: 'GMT-4', offset: 240 },
+    { name: 'GMT-3', offset: 180 },
+    { name: 'GMT-2', offset: 120 },
+    { name: 'GMT-1', offset: 60 },
+    { name: 'GMT', offset: 0 },
+    { name: 'GMT+1', offset: -60 },
+    { name: 'GMT+2', offset: -120 },
+    { name: 'GMT+3', offset: -180 },
+    { name: 'GMT+4', offset: -240 },
+    { name: 'GMT+5', offset: -300 },
+    { name: 'GMT+6', offset: -360 },
+    { name: 'GMT+7', offset: -420 },
+    { name: 'GMT+8', offset: -480 },
+    { name: 'GMT+9', offset: -540 },
+    { name: 'GMT+10', offset: -600 },
+    { name: 'GMT+11', offset: -660 },
+    { name: 'GMT+12', offset: -720 }
+];
+
+export interface DateTimePickerTimezone {
+    name: string;
+    offset: number;
+}
