+++ src/components/date-time-picker/day-view/day-view.service.ts
@@ -4,6 +4,8 @@ import { Subscription } from 'rxjs/Subscription';
 import { combineLatest } from 'rxjs/observable/combineLatest';
 import { DateTimePickerService, ModeDirection } from '../date-time-picker.service';
 import { compareDays, dateRange, gridify, months } from '../date-time-picker.utils';
+import { DateTimePickerConfig } from '../date-time-picker.config';
+import { Optional } from '@angular/core';
 
 @Injectable()
 export class DayViewService implements OnDestroy {
@@ -13,7 +15,7 @@ export class DayViewService implements OnDestroy {
 
     private _subscription: Subscription;
 
-    constructor(private _datepicker: DateTimePickerService) {
+    constructor(private _datepicker: DateTimePickerService, @Optional() private _config: DateTimePickerConfig | undefined) {
         this._subscription = combineLatest(_datepicker.month$, _datepicker.year$)
             .subscribe(([month, year]) => this.createDayGrid(month, year));
     }
@@ -32,8 +34,11 @@ export class DayViewService implements OnDestroy {
 
     private createDayGrid(month: number, year: number): void {
 
+        // get the list of months
+        const monthList = this._config ? this._config.months : months;
+
         // update the header
-        this._datepicker.setHeader(months[month] + ' ' + year);
+        this._datepicker.setHeader(monthList[month] + ' ' + year);
 
         // find the lower and upper boundaries
         const start = new Date(year, month, 1);
@@ -73,7 +78,7 @@ export class DayViewService implements OnDestroy {
 
                 // find the first day of the month
                 const first = dates.find(date => date.day === 1);
-    
+
                 // focus the date
                 this.setFocus(first.day, first.month, first.year);
             }
