+++ src/components/date-time-picker/month-view/month-view.service.ts
@@ -2,7 +2,9 @@ import { Injectable, OnDestroy } from '@angular/core';
 import { BehaviorSubject } from 'rxjs/BehaviorSubject';
 import { Subscription } from 'rxjs/Subscription';
 import { DateTimePickerService, ModeDirection } from '../date-time-picker.service';
-import { gridify, monthsShort, range } from '../date-time-picker.utils';
+import { gridify, range, monthsShort } from '../date-time-picker.utils';
+import { DateTimePickerConfig } from '../date-time-picker.config';
+import { Optional } from '@angular/core';
 
 @Injectable()
 export class MonthViewService implements OnDestroy {
@@ -12,7 +14,7 @@ export class MonthViewService implements OnDestroy {
 
     private _subscription: Subscription;
 
-    constructor(private _datepicker: DateTimePickerService) {
+    constructor(private _datepicker: DateTimePickerService, @Optional() private _config: DateTimePickerConfig | undefined) {
         this._subscription = _datepicker.year$.subscribe(year => this.createMonthGrid(year));
     }
 
@@ -43,7 +45,7 @@ export class MonthViewService implements OnDestroy {
         // create a 4x3 grid of month numbers
         const months: MonthViewItem[] = range(0, 11).map(month => {
             return {
-                name: monthsShort[month],
+                name: this._config ? this._config.monthsShort[month] : monthsShort[month],
                 month: month,
                 year: year,
                 isCurrentMonth: year === currentYear && month === currentMonth,
@@ -59,7 +61,7 @@ export class MonthViewService implements OnDestroy {
 
         // if there is no focused month select the first one
         if (this._datepicker.modeDirection === ModeDirection.Descend && this.focused$.value === null) {
-            
+
             // check if the selected month is in view
             const selectedMonth = months.find(month => month.isActiveMonth);
